refactors tests into more manageable structure
