--- test/integration/owner_can_view_rentals_test.rb
@@ -1,16 +1,0 @@
-require 'test_helper'
-
-class OwnerCanViewRentalsTest < ActionDispatch::IntegrationTest
-  test "owner can view all items" do
-    skip
-    create_rentals(2, "Castle")
-    login_owner
-    click_link ("View All Rentals")
-
-    assert "/owners/rentals", current_path
-    assert page.has_content?("Rentals")
-    assert page.has_content?("Castle")
-    assert page.has_content?("ACTIVE")
-    assert page.has_button?("Edit")
-  end
-end
